Rename ActionsList file to index
This simplifies the code a bit
