+++ assets/src/blocks/editorIndex.js
@@ -1,5 +1,5 @@
 
-import {registerActionsList} from './ActionsList/ActionsList';
+import {registerActionsList} from './ActionsList';
 
 // Beta blocks
 registerActionsList();
